아이템 생성 매크로 추가, 클라이언트 체력 추가, 아이템 강화수치 추가, 캐릭터 갱신 패킷 추가, 기본 유물 장착 추가
아이템 생성 매크로 추가
클라이언트 체력 추가
아이템 강화수치 추가
캐릭터 갱신 패킷 추가
기본 유물 장착 추가

+++ Server-Library/Header/MemoryStream.h
@@ -16,7 +16,7 @@ public:
 	{
 	}
 
-	OutputMemoryStream(int capacity)
+	OutputMemoryStream(size_t capacity)
 		: mBuffer(nullptr)
 		, mHead(0)
 		, mCapacity(capacity)
